Further cleanups.
