+++ test/CMakeLists.txt
@@ -119,6 +119,7 @@ add_subdirectory(engine)
 add_subdirectory(parser)
 add_subdirectory(index)
 add_subdirectory(backports)
+add_subdirectory(rdfTypes)
 
 addLinkAndDiscoverTest(ValueIdComparatorsTest util)
 
@@ -244,8 +245,6 @@ addLinkAndDiscoverTest(StreamableBodyTest http)
 addLinkAndDiscoverTest(StreamableGeneratorTest)
 
 addLinkAndDiscoverTest(AcceptHeaderTest mediaTypes httpParser)
-
-addLinkAndDiscoverTest(RdfEscapingTest parser)
 
 addLinkAndDiscoverTestNoLibs(CompactStringVectorTest)
