+++ test/rdfTypes/CMakeLists.txt
@@ -1,0 +1,2 @@
+addLinkAndDiscoverTestNoLibs(VariableTest rdfTypes)
+addLinkAndDiscoverTestNoLibs(RdfEscapingTest rdfTypes)
